Conflicts:
	Tst/New.lst
	kernel/GBEngine/kstd2.cc
	kernel/GBEngine/kutil.cc

+++ libpolys/coeffs/coeffs.h
@@ -181,8 +181,8 @@ struct n_Procs_s
    /// how complicated, (0) => 0, or positive
    int     (*cfSize)(number n, const coeffs r);
 
-   /// convertion to int, 0 if impossible
-   int     (*cfInt)(number &n, const coeffs r);
+   /// convertion to long, 0 if impossible
+   long    (*cfInt)(number &n, const coeffs r);
 
    /// Converts a non-negative number n into a GMP number, 0 if impossible
    void     (*cfMPZ)(mpz_t result, number &n, const coeffs r);
@@ -543,7 +543,7 @@ static FORCE_INLINE number n_InitMPZ(mpz_t n,     const coeffs r)
 
 /// conversion of n to an int; 0 if not possible
 /// in Z/pZ: the representing int lying in (-p/2 .. p/2]
-static FORCE_INLINE int n_Int(number &n,       const coeffs r)
+static FORCE_INLINE long n_Int(number &n,       const coeffs r)
 { STATISTIC(n_Int); assume(r != NULL); assume(r->cfInt!=NULL); return r->cfInt(n,r); }
 
 /// conversion of n to a GMP integer; 0 if not possible
@@ -608,7 +608,8 @@ static FORCE_INLINE number n_GetNumerator(number& n, const coeffs r)
 { STATISTIC(n_GetNumerator); assume(r != NULL); assume(r->cfGetNumerator!=NULL); return r->cfGetNumerator(n, r); }
 
 /// return the quotient of 'a' and 'b', i.e., a/b;
-/// raise an error if 'b' is not invertible in r
+/// raises an error if 'b' is not invertible in r
+/// exception in Z: raises an error if 'a' is not divisible by 'b'
 static FORCE_INLINE number n_Div(number a, number b, const coeffs r)
 { STATISTIC(n_Div); assume(r != NULL); assume(r->cfDiv!=NULL); return r->cfDiv(a,b,r); }
